feat: types 디렉토리 분리
