+++ view/rbgksqkr/src/components/AxisBottom.tsx
@@ -1,10 +1,6 @@
-import { ScaleBand, select, axisBottom } from "d3";
+import { select, axisBottom } from "d3";
 import { useEffect, useRef } from "react";
-
-interface AxisBottomProps {
-    scale: ScaleBand<string>;
-    transform: string;
-}
+import { AxisBottomProps } from "../types/bar";
 
 const AxisBottom = ({ scale, transform }: AxisBottomProps) => {
     const ref = useRef<SVGGElement>(null);
