+++ view/rbgksqkr/src/sampleData.ts
@@ -1,3 +1,5 @@
+import { Data } from "./types/bar";
+
 export const lineData = [
     { d: "2007-04-23", v: 93.24 },
     { d: "2007-04-24", v: 95.35 },
@@ -1280,10 +1282,6 @@ export const lineData = [
     { d: "2012-04-29", v: 583.98 },
     { d: "2012-05-01", v: 582.13 },
 ];
-export interface Data {
-    label: string;
-    value: number;
-}
 
 export const barData: Data[] = [
     { label: "Apples", value: 100 },
